getFetch module
