+++ packages/client/src/index.ts
@@ -1,3 +1,4 @@
 export * from './createTRPCClient';
+export * from './getFetch';
 export * from './links/core';
 export * from './TRPCClientError';
