+++ packages/client/src/internals/fetchHelpers.ts
@@ -1,4 +1,4 @@
-function getWindow() {
+export function getWindow() {
   if (typeof window !== 'undefined') {
     return window;
   }
@@ -9,17 +9,4 @@ export function getAbortController(
 ): typeof AbortController | null {
   const win = getWindow();
   return ac || win.AbortController || null;
-}
-export function getFetch(f?: typeof fetch): typeof fetch {
-  if (f) {
-    return f;
-  }
-  const win = getWindow();
-  if (win.fetch) {
-    return typeof win.fetch.bind === 'function'
-      ? win.fetch.bind(win)
-      : win.fetch;
-  }
-
-  throw new Error('No fetch implementation found');
 }
